+++ src/main/java/pretzel/dreamketcherbe/domain/auth/google/GoogleApiClient.java
@@ -1,4 +1,4 @@
-package pretzel.dreamketcherbe.auth.google;
+package pretzel.dreamketcherbe.domain.auth.google;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.util.MultiValueMap;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.PostExchange;
-import pretzel.dreamketcherbe.auth.google.dto.GoogleToken;
-import pretzel.dreamketcherbe.auth.google.dto.GoogleUserInfo;
+import pretzel.dreamketcherbe.domain.auth.google.dto.GoogleToken;
+import pretzel.dreamketcherbe.domain.auth.google.dto.GoogleUserInfo;
 
 public interface GoogleApiClient {
