+++ src/main/java/pretzel/dreamketcherbe/domain/auth/service/internal/AuthContext.java
@@ -1,10 +1,10 @@
-package pretzel.dreamketcherbe.auth.service.internal;
+package pretzel.dreamketcherbe.domain.auth.service.internal;
 
 
 import org.springframework.stereotype.Component;;
 import org.springframework.web.context.annotation.RequestScope;
-import pretzel.dreamketcherbe.auth.exception.AuthException;
-import pretzel.dreamketcherbe.auth.exception.AuthExceptionType;
+import pretzel.dreamketcherbe.domain.auth.exception.AuthException;
+import pretzel.dreamketcherbe.domain.auth.exception.AuthExceptionType;
 
 @Component
 @RequestScope
